Added track bar for heading, changed some visual aspects.
Added a track bar for the Street View to change the heading. Each
position in the track bar changes the heading by 45 degrees. Also
changed that whenever you scroll on either track bar, you do not need to
press the "Static Map" or the "Street View" button. The map changes
automatically. Also changed the color of some buttons and made labels
transparent.